pyproject.toml
[project]
name = "adgen-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "dotenv>=0.9.9",
    "google-genai>=1.51.0",
    "google-search-results>=2.4.2",
    "openai>=2.8.1",
    "pillow>=12.0.0",
    "pinscrape>=5.0.0",
    "streamlit>=1.28.0",
]
